AP_NavEKF3: Avoid switching EKF3 right after yaw, pos or vel resets
After resets on a badly estimating EKF lane, the error score may be unrealistically
low for a short while, which could trigger a lane switch. This logic should avoid
switching based on error score alone for a period of 1 second.
If the primary EKF is unheathy, this logic will not prevent switching.
The recent_reset flag is stored in the solution status bitmask, so that it will also
be logged on the XKF4 log message.